+++ pages/company/components/company-wizard-nav.tsx
@@ -1,6 +1,5 @@
 import { useCallback } from 'react';
 import {
-  Heading,
   RouterLink,
   WizardNavigation,
   WizardNavigationItem,
@@ -12,63 +11,83 @@ import { Step } from '@/context/company-context';
 
 const COMPANY_STEPS = [
   {
-    label: '1.1 - Registrant',
+    label: '1. Registrant',
     step: 'company.registrant' as Step,
   },
   {
-    label: '1.2 - Company details',
+    label: '2. Company details',
     step: 'company.companyDetails' as Step,
   },
   {
-    label: '1.3 - Company address',
+    label: '3. Company address',
     step: 'company.companyAddress' as Step,
   },
   {
-    label: '1.4 - Share series',
+    label: '4. Share series',
     step: 'company.shareSeries' as Step,
   },
   {
-    label: '1.5 - Managing directors',
+    label: '5. Managing directors',
     step: 'company.managingDirectors' as Step,
   },
   {
-    label: '1.6 - Board members',
+    label: '6. Board members',
     step: 'company.boardMembers' as Step,
   },
   {
-    label: '1.7 - Auditor',
-    step: 'company.auditor' as Step,
+    label: '7. Auditor',
+    step: 'company.auditorDetails' as Step,
   },
 ];
 
 const BENEFICIAL_OWNER_STEPS = [
-  { label: '2.1 - Share series', step: 'beneficialOwners.shareSeries' as Step },
+  { label: '1. Share series', step: 'beneficialOwners.shareSeries' as Step },
   {
-    label: '2.2 - Shareholders',
+    label: '2. Shareholders',
     step: 'beneficialOwners.shareholders' as Step,
   },
 ];
 
 const SIGNATORY_RIGHTS_STEPS = [
   {
-    label: '3.1 - Signin rights',
+    label: '1. Signin rights',
     step: 'signatoryRights.signinRights' as Step,
   },
 ];
 
 interface Props {
+  wizardType: 'company' | 'beneficialOwners' | 'signatoryRights';
   onWizardNavChange: (clickedStep: number) => void;
 }
 
 export default function CompanyWizardNav(props: Props) {
-  const { onWizardNavChange } = props;
+  const { wizardType, onWizardNavChange } = props;
   const { width } = useDimensions();
   const { isStepDone, isPrevStepDone, doneSteps, step, isLoading } =
     useCompanyContext();
 
-  const doneStepValues = Object.values(doneSteps);
+  const navHeading =
+    wizardType === 'company'
+      ? 'Details'
+      : wizardType === 'beneficialOwners'
+      ? 'Beneficial owners'
+      : 'Signatory rights';
+  const navSteps =
+    wizardType === 'company'
+      ? COMPANY_STEPS
+      : wizardType === 'beneficialOwners'
+      ? BENEFICIAL_OWNER_STEPS
+      : SIGNATORY_RIGHTS_STEPS;
+
+  const trackedDoneSteps = Object.keys(doneSteps)
+    .filter(key => key.includes(wizardType))
+    .reduce((cur, key) => {
+      return Object.assign(cur, { [key]: doneSteps[key] });
+    }, {});
+  const doneStepValues = Object.values(trackedDoneSteps);
   const allStepsDone =
-    doneStepValues.length === 10 && doneStepValues.every(isDone => isDone);
+    doneStepValues.length === navSteps.length &&
+    doneStepValues.every(isDone => isDone);
 
   const getItemStatus = useCallback(
     (
@@ -99,16 +118,13 @@ export default function CompanyWizardNav(props: Props) {
 
   return (
     <WizardNavigation
-      heading="Stages"
+      heading={navHeading}
       aria-label="Company establishment wizard steps"
       initiallyExpanded={false}
       variant={width > 1024 ? 'default' : 'smallScreen'}
       className={isLoading ? 'pointer-events-none' : ''}
     >
-      <div className="ml-5 mb-2">
-        <Heading variant="h5">1. Details</Heading>
-      </div>
-      {COMPANY_STEPS.map((item, index) => {
+      {navSteps.map((item, index) => {
         const status = getItemStatus(item.step, index, step);
 
         return (
@@ -119,47 +135,17 @@ export default function CompanyWizardNav(props: Props) {
           </WizardNavigationItem>
         );
       })}
-
-      <div className="ml-5 mb-2 mt-4">
-        <Heading variant="h5">2. Beneficial owners</Heading>
-      </div>
-      {BENEFICIAL_OWNER_STEPS.map((item, i) => {
-        const index = i + 7;
-        const status = getItemStatus(item.step, index, step);
-
-        return (
-          <WizardNavigationItem key={item.step} status={status}>
-            <RouterLink onClick={() => onWizardNavChange(index)}>
-              {item.label}
-            </RouterLink>
-          </WizardNavigationItem>
-        );
-      })}
-
-      <div className="ml-5 mb-2 mt-4">
-        <Heading variant="h5">3. Signatory rights</Heading>
-      </div>
-      {SIGNATORY_RIGHTS_STEPS.map((item, i) => {
-        const index = i + 9;
-        const status = getItemStatus(item.step, index, step);
-
-        return (
-          <WizardNavigationItem key={item.step} status={status}>
-            <RouterLink onClick={() => onWizardNavChange(index)}>
-              {item.label}
-            </RouterLink>
-          </WizardNavigationItem>
-        );
-      })}
-
-      <div className="ml-5 mb-2 mt-4">
-        <Heading variant="h5">4. Preview</Heading>
-      </div>
       <WizardNavigationItem
-        status={allStepsDone ? (step === 10 ? 'current' : 'default') : 'coming'}
+        status={
+          allStepsDone
+            ? step === navSteps.length
+              ? 'current-completed'
+              : 'default'
+            : 'coming'
+        }
       >
-        <RouterLink onClick={() => onWizardNavChange(10)}>
-          4.1 Preview and submit
+        <RouterLink onClick={() => onWizardNavChange(navSteps.length)}>
+          {`${navSteps.length + 1}.`} Preview
         </RouterLink>
       </WizardNavigationItem>
     </WizardNavigation>
