+++ pages/company/establishment/details.page.tsx
@@ -1,7 +1,7 @@
 import { CompanyContextProvider } from '@/context/company-context';
 import AuthSentry from '@/components/auth-sentry';
 import Page from '@/components/layout/page';
-import CompanyWizard from '../components/company-wizard-edit';
+import CompanyWizard from '../components/company-wizard';
 
 export default function DetailsPage() {
   return (
